ADD: backbone to handle both Gurobi & OR-Tools solvers
